+++ __test__/inputs/HtmlEditor.spec.js
@@ -1,7 +1,7 @@
 import chai from "chai";// eslint-disable-line import/no-extraneous-dependencies
 import React from "react";
 import ReactDOM from "react-dom";
-import TestUtils from "react-addons-test-utils";// eslint-disable-line import/no-extraneous-dependencies
+import TestUtils from "../TestUtils";
 import HtmlEditor from "inputs/htmleditor/HtmlEditor";// eslint-disable-line import/no-extraneous-dependencies,import/no-unresolved
 
 describe("inputs/htmleditor/HtmlEditor", () => {