+++ __test__/notification/Notification.spec.js
@@ -1,5 +1,6 @@
 import chai from "chai"; // eslint-disable-line import/no-extraneous-dependencies
 import React from "react";
+import TestUtils from "../TestUtils";
 import Notification from "notification/Notification"; // eslint-disable-line import/no-extraneous-dependencies,import/no-unresolved
 import NotificationItem from "notification/NotificationItem"; // eslint-disable-line import/no-extraneous-dependencies,import/no-unresolved
 import { mount } from "enzyme"; // eslint-disable-line import/no-extraneous-dependencies