+++ __test__/wizard/Wizard.spec.js
@@ -1,7 +1,8 @@
 import chai from "chai";// eslint-disable-line import/no-extraneous-dependencies
 import React from "react";
-import Wizard from "wizard/Wizard";// eslint-disable-line import/no-extraneous-dependencies,import/no-unresolved
 import TestUtils from "../TestUtils";
+import Wizard from "wizard/Wizard";// eslint-disable-line import/no-extraneous-dependencies,import/no-unresolved
+
 
 describe("wizard/Wizard", () => {
    //todo